requirements.txt
aenum==3.1.11
antlr4-python3-runtime==4.9.3
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.8
attrs==22.1.0
backcall==0.2.0
bddl==1.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
blis==0.7.8
catalogue==2.0.8
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
cloudpickle==2.1.0
cmake==3.24.1
coloredlogs==15.0.1
cycler==0.11.0
cymem==2.0.6
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.5.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.4.1/en_core_web_sm-3.4.1-py3-none-any.whl
entrypoints==0.4
executing==0.10.0
fastjsonschema==2.16.1
filelock==3.8.0
fonttools==4.33.3
freetype-py==2.3.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
GPUtil==1.4.0
gym==0.21.0
gym-minigrid==1.0.3
h5py==3.7.0
huggingface-hub==0.10.1
humanfriendly==10.0
hydra-core==1.2.0
idna==3.3
iniconfig==1.1.1
ipykernel==6.15.1
ipython==8.4.0
ipython-genutils==0.2.0
ipywidgets==8.0.1
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.13.0
jupyter==1.0.0
jupyter-client==7.3.4
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.2
jupytext==1.14.1
kiwisolver==1.4.3
langcodes==3.3.0
lxml==4.9.1
markdown-it-py==2.1.0
MarkupSafe==2.1.1
matplotlib==3.5.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.0
mdurl==0.1.2
mistune==0.8.4
murmurhash==1.0.8
nbclient==0.6.6
nbconvert==6.5.3
nbformat==5.4.0
nest-asyncio==1.5.5
networkx==2.8.5
nltk==3.7
notebook==6.4.12
numpy==1.23.0
omegaconf==2.2.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pandocfilters==1.5.0
parso==0.8.3
pathy==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.2.0
pluggy==1.0.0
preshed==3.0.7
progressbar==2.5
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==3.20.1
psutil==5.9.1
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
py360convert==0.1.0
pybullet-svl==3.1.6.4
pycparser==2.21
pydantic==1.9.2
Pygments==2.13.0
pyinstrument==4.2.0
pyparsing==3.0.9
pyrsistent==0.18.1
pytest==7.1.2
python-dateutil==2.8.2
pytz==2022.2.1
PyYAML==6.0
pyzmq==23.2.1
qtconsole==5.3.1
QtPy==2.2.0
regex==2022.9.13
requests==2.28.1
scikit-learn==1.1.3
scipy==1.9.0
Send2Trash==1.8.0
sentence-transformers==2.2.2
sentencepiece==0.1.97
six==1.16.0
smart-open==5.2.1
smmap==5.0.0
soupsieve==2.3.2.post1
spacy==3.4.1
spacy-legacy==3.0.10
spacy-loggers==1.0.3
srsly==2.4.4
stable-baselines3==1.6.0
stack-data==0.4.0
tensorboardX==2.5.1
terminado==0.15.0
thinc==8.1.0
threadpoolctl==3.1.0
tinycss2==1.1.1
tokenizers==0.13.1
toml==0.10.2
tomli==2.0.1
tornado==6.2
tqdm==4.64.0
traitlets==5.3.0
transformers==4.23.1
transforms3d==0.3.1
trimesh==3.13.5
typer==0.4.2
typing_extensions==4.3.0
urllib3==1.26.11
wasabi==0.10.1
wcwidth==0.2.5
webencodings==0.5.1
widgetsnbextension==4.0.2